pyproject.toml
[project]
name = "pruebadjango"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.0.1",
]
